Split Buttons into hardware and dispatch
